+++ openghg/analyse/__init__.py
@@ -1,4 +1,5 @@
 from ._alignment import combine_datasets
+from ._fp_x_flux import fp_x_flux_keep_space, warm_numba_fp_x_flux, write_fp_x_flux_keep_space_zarr
 from ._modelled_obs import fp_x_flux_integrated, make_integrated_low_freq_flux
 from ._scenario import ModelScenario
 from ._utils import (